% Auto-generated: do not edit by hand
\name{'dam'HeaderMenu}

\alias{'dam'HeaderMenu}

\title{HeaderMenu component}

\description{
MenuHeader element to display in a chat, adds icon for resizing the window and user options such as resetting the chat  @component @returns {JSX.Element} Rendered ChatWindow component.
}

\usage{
'dam'HeaderMenu(id=NULL, headerText=NULL, resetClicked=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{headerText}{Character. What text to display in the header.}

\item{resetClicked}{Logical. If the reset button was clicked or not}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

