% Auto-generated: do not edit by hand
\name{'dam'ResizableWindow}

\alias{'dam'ResizableWindow}

\title{ResizableWindow component}

\description{
ResizableWindow displays a resizable container for chat content.  @component @param {Object} props - Component props. @param {string} props.id - The ID used to identify this component in Dash callbacks. @param {Function} props.setProps - Dash-assigned callback to report property changes. @returns {JSX.Element} A resizable chat container.
}

\usage{
'dam'ResizableWindow(children=NULL, id=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Child component to display in the footer.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

